feat: reorganize timeline view
